requirements.txt
# Core deep learning frameworks
torch>=1.9.0
torch-geometric>=2.0.0
torch-scatter>=2.0.8
torch-sparse>=0.6.12
torch-cluster>=1.5.9
torch-spline-conv>=1.2.1

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data processing
h5py>=3.1.0
nibabel>=3.2.0
nilearn>=0.8.0
biom-format>=2.1.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
networkx>=2.6.0

# Configuration and utilities
pyyaml>=5.4.0
tqdm>=4.62.0
wandb>=0.12.0
tensorboard>=2.7.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Optional: SHAP for interpretability
shap>=0.40.0

# Optional: Additional visualization
bokeh>=2.4.0
holoviews>=1.14.0 